+++ packages/sdk/src/Actions/EventAction.test.ts
@@ -1,13 +1,16 @@
 import { selectors as eventSelectors } from '@boostxyz/signatures/events';
+import { selectors as funcSelectors } from '@boostxyz/signatures/functions';
 import { loadFixture } from '@nomicfoundation/hardhat-network-helpers';
-import { type Hex, type Log, isAddress } from 'viem';
+import { type Hex, type Log, isAddress, pad, parseEther, toHex } from 'viem';
 import { beforeAll, beforeEach, describe, expect, test } from 'vitest';
+import type { MockERC20 } from '../../test/MockERC20';
 import type { MockERC721 } from '../../test/MockERC721';
 import { accounts } from '../../test/accounts';
 import {
   type Fixtures,
   defaultOptions,
   deployFixtures,
+  fundErc20,
   fundErc721,
 } from '../../test/helpers';
 import {
@@ -18,7 +21,7 @@ import {
   SignatureType,
 } from './EventAction';
 
-let fixtures: Fixtures, erc721: MockERC721;
+let fixtures: Fixtures, erc721: MockERC721, erc20: MockERC20;
 
 beforeAll(async () => {
   fixtures = await loadFixture(deployFixtures);
@@ -52,48 +55,98 @@ function basicErc721TransferAction(
   };
 }
 
-/*
+function cloneEventAction(fixtures: Fixtures, erc721: MockERC721) {
+  return function cloneEventAction() {
+    return fixtures.registry.initialize(
+      crypto.randomUUID(),
+      fixtures.core.EventAction(basicErc721TransferAction(erc721)),
+    );
+  };
+}
+
 function basicErc721MintFuncAction(
   erc721: MockERC721,
 ): EventActionPayloadSimple {
   return {
     actionClaimant: {
       signatureType: SignatureType.FUNC,
-      signature: funcSelectors['mint(address)'] as Hex,
+      signature: pad(funcSelectors['mint(address)'] as Hex),
       fieldIndex: 0,
       targetContract: erc721.assertValidAddress(),
+      chainid: defaultOptions.config.chains[0].id,
     },
     actionSteps: [
       {
-        signature: funcSelectors['mint(address)'] as Hex,
+        signature: pad(funcSelectors['mint(address)'] as Hex),
         signatureType: SignatureType.FUNC,
         actionType: 0,
         targetContract: erc721.assertValidAddress(),
+        chainid: defaultOptions.config.chains[0].id,
         actionParameter: {
           filterType: FilterType.EQUAL,
           fieldType: PrimitiveType.ADDRESS,
-          fieldIndex: 2,
+          fieldIndex: 0,
           filterData: accounts[1].account,
         },
       },
     ],
   };
 }
-*/
 
-function cloneEventAction(fixtures: Fixtures, erc721: MockERC721) {
-  return function cloneEventAction() {
+function basicErc20MintFuncAction(erc20: MockERC20): EventActionPayloadSimple {
+  return {
+    actionClaimant: {
+      signatureType: SignatureType.FUNC,
+      signature: pad(funcSelectors['mint(address to, uint256 amount)'] as Hex),
+      fieldIndex: 0,
+      targetContract: erc20.assertValidAddress(),
+      chainid: defaultOptions.config.chains[0].id,
+    },
+    actionSteps: [
+      {
+        signature: pad(
+          funcSelectors['mint(address to, uint256 amount)'] as Hex,
+        ),
+        signatureType: SignatureType.FUNC,
+        actionType: 0,
+        targetContract: erc20.assertValidAddress(),
+        chainid: defaultOptions.config.chains[0].id,
+        actionParameter: {
+          filterType: FilterType.EQUAL,
+          fieldType: PrimitiveType.ADDRESS,
+          fieldIndex: 0,
+          filterData: accounts[1].account,
+        },
+      },
+    ],
+  };
+}
+
+function cloneFunctionAction20(fixtures: Fixtures, erc20: MockERC20) {
+  return function cloneFunctionAction20() {
     return fixtures.registry.clone(
       crypto.randomUUID(),
       new fixtures.bases.EventAction(
         defaultOptions,
-        basicErc721TransferAction(erc721),
+        basicErc20MintFuncAction(erc20),
       ),
     );
   };
 }
 
-describe('EventAction', () => {
+function cloneFunctionAction(fixtures: Fixtures, erc721: MockERC721) {
+  return function cloneFunctionAction() {
+    return fixtures.registry.clone(
+      crypto.randomUUID(),
+      new fixtures.bases.EventAction(
+        defaultOptions,
+        basicErc721MintFuncAction(erc721),
+      ),
+    );
+  };
+}
+
+describe('EventAction Event Selector', () => {
   beforeEach(async () => {
     erc721 = await loadFixture(fundErc721(defaultOptions));
   });
@@ -204,5 +257,162 @@ describe('EventAction', () => {
     } as Log;
     const action = await loadFixture(cloneEventAction(fixtures, erc721));
     expect(await action.validateActionSteps({ logs: [log] })).toBe(true);
+  });
+});
+
+describe('EventAction Func Selector', () => {
+  beforeEach(async () => {
+    erc721 = await loadFixture(fundErc721(defaultOptions));
+    erc20 = await loadFixture(fundErc20(defaultOptions));
+  });
+
+  test('can be deployed successfully', async () => {
+    const action = new EventAction(
+      defaultOptions,
+      basicErc721MintFuncAction(erc721),
+    );
+    await action.deploy();
+    expect(isAddress(action.assertValidAddress())).toBe(true);
+  });
+
+  test('validates function action step with correct hash', async () => {
+    const action = await loadFixture(cloneFunctionAction(fixtures, erc721));
+    const actionSteps = await action.getActionSteps();
+    const recipient = accounts[1].account;
+    const { hash } = await erc721.mintRaw(recipient, {
+      value: parseEther('.1'),
+    });
+
+    expect(
+      await action.isActionFunctionValid(actionSteps[0], {
+        hash,
+      }),
+    ).toBe(true);
+  });
+
+  test('throws an error when hash is missing', async () => {
+    const action = await loadFixture(cloneFunctionAction(fixtures, erc721));
+    const actionSteps = await action.getActionSteps();
+    try {
+      await action.isActionFunctionValid(actionSteps[0], {});
+    } catch (e) {
+      expect(e).toBeInstanceOf(Error);
+      expect((e as Error).message).toBe(
+        'Hash is required for function validation',
+      );
+    }
+  });
+
+  test('validates function step with EQUAL filter', async () => {
+    const action = await loadFixture(cloneFunctionAction(fixtures, erc721));
+    const actionSteps = await action.getActionSteps();
+
+    const recipient = accounts[1].account;
+    const { hash } = await erc721.mintRaw(recipient, {
+      value: parseEther('.1'),
+    });
+
+    const criteriaMatch = await action.isActionFunctionValid(actionSteps[0], {
+      hash,
+    });
+
+    expect(criteriaMatch).toBe(true);
+  });
+
+  test('fails validation with incorrect function signature', async () => {
+    const action = await loadFixture(cloneFunctionAction(fixtures, erc721));
+    const actionSteps = await action.getActionSteps();
+    const recipient = accounts[1].account;
+
+    const invalidStep = {
+      ...actionSteps[0],
+      signature: pad(funcSelectors['mint(address to, uint256 amount)'] as Hex), // Intentional mismatch
+    };
+
+    const { hash } = await erc721.mintRaw(recipient, {
+      value: parseEther('.1'),
+    });
+
+    try {
+      await action.isActionFunctionValid(invalidStep, { hash });
+    } catch (e) {
+      expect(e).toBeInstanceOf(Error);
+      expect((e as Error).message).toContain(
+        'Failed to decode function data: Encoded function signature "0x6a627842"',
+      );
+    }
+  });
+
+  test('validates against NOT_EQUAL filter criteria', async () => {
+    const action = await loadFixture(cloneFunctionAction(fixtures, erc721));
+    const actionSteps = await action.getActionSteps();
+    actionSteps[0].actionParameter.filterType = FilterType.NOT_EQUAL;
+    const recipient = accounts[2].account;
+    const { hash } = await erc721.mintRaw(recipient, {
+      value: parseEther('.1'),
+    });
+
+    expect(
+      await action.isActionFunctionValid(actionSteps[0], {
+        hash,
+      }),
+    ).toBe(true);
+  });
+
+  test('validates GREATER_THAN criteria for numeric values', async () => {
+    const action = await loadFixture(cloneFunctionAction20(fixtures, erc20));
+    const actionSteps = await action.getActionSteps();
+
+    actionSteps[0].actionParameter = {
+      filterType: FilterType.GREATER_THAN,
+      fieldType: PrimitiveType.UINT,
+      fieldIndex: 1,
+      filterData: toHex('1'),
+    };
+
+    const address = accounts[1].account;
+    const value = 400n;
+    const { hash } = await erc20.mintRaw(address, value);
+
+    expect(
+      await action.isActionFunctionValid(actionSteps[0], {
+        hash,
+      }),
+    ).toBe(true);
+  });
+
+  test('validates LESS_THAN criteria for numeric values', async () => {
+    const action = await loadFixture(cloneFunctionAction20(fixtures, erc20));
+    const actionSteps = await action.getActionSteps();
+    actionSteps[0].actionParameter = {
+      filterType: FilterType.LESS_THAN,
+      fieldType: PrimitiveType.UINT,
+      fieldIndex: 1,
+      filterData: toHex('5'),
+    };
+
+    const address = accounts[1].account;
+    const value = 4n;
+    const { hash } = await erc20.mintRaw(address, value);
+
+    expect(
+      await action.isActionFunctionValid(actionSteps[0], {
+        hash,
+      }),
+    ).toBe(true);
+  });
+
+  test('validates entire flow of function action', async () => {
+    const action = await loadFixture(cloneFunctionAction(fixtures, erc721));
+    const recipient = accounts[1].account;
+    const { hash } = await erc721.mintRaw(recipient, {
+      value: parseEther('.1'),
+    });
+
+    expect(
+      await action.validateActionSteps({
+        hash,
+      }),
+    ).toBe(true);
   });
 });