+++ packages/sdk/src/AllowLists/SimpleAllowList.test.ts
@@ -16,9 +16,9 @@ beforeAll(async () => {
 
 function freshAllowList(fixtures: Fixtures) {
   return function freshAllowList() {
-    return fixtures.registry.clone(
+    return fixtures.registry.initialize(
       crypto.randomUUID(),
-      new fixtures.bases.SimpleAllowList(defaultOptions, {
+      fixtures.core.SimpleAllowList({
         owner: defaultOptions.account.address,
         allowed: [defaultOptions.account.address],
       }),