+++ packages/sdk/src/Incentives/AllowListIncentive.test.ts
@@ -15,9 +15,9 @@ let fixtures: Fixtures;
 
 function freshAllowList(fixtures: Fixtures) {
   return function freshAllowList() {
-    return fixtures.registry.clone(
+    return fixtures.registry.initialize(
       crypto.randomUUID(),
-      new fixtures.bases.SimpleAllowList(defaultOptions, {
+      fixtures.core.SimpleAllowList({
         owner: defaultOptions.account.address,
         allowed: [],
       }),