+++ packages/signatures/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @boostxyz/signatures
 
+## 0.0.0-alpha.2
+
+### Minor Changes
+
+- 9e9757f: add NameRegistered event to signatures
+
 ## 0.0.0-alpha.1
 
 ### Minor Changes